Move USUM tutor moves to their own editor (#374)
Splits the functionality from the existing g7 Mart cro editor to a separate form.
